Camera Controller and Camera Transitons
I added two scripts. One is the Camera Controlller, which makes sure the camera moves to wherever it needs to be. The other is the Camera Transition script; when the ball hits a trigger with this script, it will update the camera's settings.

Note that the camera has to be on 0 in the z-axis and in orthagraphic mode to work. If we need a perspective camera, I can change it.